fix(tofu): remove lalt tap dance
When trying to alt tab finding that the tab key was intrupting the dance
triggering TD_SINGLE_TAP. This does not trigger the alt but the one shot
layer key.

Tried to add a new TD_SINGLE_INTERUPT_TAP but found that I had to wait
for the tapping term for other keys like the number/function keys.

Need a better solution in the future.