feat(qmk): replace Achordion with Chordal Hold
This commit enables the use of the new built-in Chordal Hold tap-hold
option. This update is currently in `develop` branch of qmk and has not
made it into the main qmk branch yet.

Achordion is still enabled in my other older keymaps and in the
userspace utility, however it is not enabled by default in the user's
rules.mk file.